+++ nre/libs/libstdc++/arch/ExecEnv.cc
@@ -55,7 +55,7 @@ void ExecEnv::thread_exit() {
 
 void *ExecEnv::setup_stack(Pd *pd,Thread *t,startup_func start,uintptr_t ret,uintptr_t stack) {
 	void **sp = reinterpret_cast<void**>(stack);
-	unsigned stack_top = STACK_SIZE / sizeof(void*);
+	size_t stack_top = STACK_SIZE / sizeof(void*);
 	sp[--stack_top] = t;
 	sp[--stack_top] = pd;
 	sp[--stack_top] = reinterpret_cast<void*>(start);
